[TIMOB-20251]: Move storage permission to Filesystem. No caching if permission is not enabled.
