+++ android/titanium/src/java/org/appcelerator/titanium/util/TiResponseCache.java
@@ -311,7 +311,7 @@ public class TiResponseCache extends ResponseCache
 	public CacheResponse get(URI uri, String rqstMethod,
 			Map<String, List<String>> rqstHeaders) throws IOException 
 	{
-		if (uri == null || cacheDir == null) return null;
+		if (!TiFileHelper2.hasStoragePermission() || uri == null || cacheDir == null) return null;
 
 		// Workaround for https://jira.appcelerator.org/browse/TIMOB-18913
 		// This workaround should be removed when HTTPClient is refactored with HttpUrlConnection
@@ -408,7 +408,7 @@ public class TiResponseCache extends ResponseCache
 	@Override
 	public CacheRequest put(URI uri, URLConnection conn) throws IOException
 	{
-		if (cacheDir == null) return null;
+		if (cacheDir == null || !TiFileHelper2.hasStoragePermission()) return null;
 
 		// Workaround for https://jira.appcelerator.org/browse/TIMOB-18913
 		// This workaround should be removed when HTTPClient is refactored with HttpUrlConnection
