+++ apidoc/Titanium/Filesystem/Filesystem.yml
@@ -119,6 +119,24 @@ methods:
     returns:
         type: Boolean
 
+  - name: hasStoragePermissions
+    summary: Returns `true` if the app has storage permissions.
+    returns:
+        type: Boolean
+    since: "6.0.0"
+
+  - name: requestStoragePermissions
+    summary: Requests for storage permissions
+    description: |
+        On Android, the request view will show if the permission is not accepted by the user, and the user did
+        not check the box "Never ask again" when denying the request. If the user checks the box "Never ask again,"
+        the user has to manually enable the permission in device settings.
+    parameters:
+      - name: callback
+        summary: Function to call upon user decision to grant storage access
+        type: Callback<RequestStorageAccessResult>
+    since: "6.0.0"
+
   - name: openStream
     summary: Opens file using the <Titanium.IOStream> interface.
     parameters:
@@ -341,3 +359,9 @@ properties:
         may not persist when the application is shut down and restarted.
     type: String
     permission: read-only
+    
+---
+name: RequestStorageAccessResult
+summary: Argument passed to the callback when a request finishes successfully or erroneously.
+extends: ErrorResponse
+
